jumboSmash/requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
Django==3.0.2
djangorestframework==3.11.0
kombu==4.6.7
mock==4.0.1
pytz==2019.3
sqlparse==0.3.0
celery==4.4.0
